+++ lib/chewy/query/loading.rb
@@ -12,15 +12,11 @@ module Chewy
 
       def _load_objects(options)
         loaded_objects = Hash[_results.group_by(&:class).map do |type, objects|
-          model = type.adapter.model
-          scope = model.where(id: objects.map(&:id))
-          additional_scope = options[:scopes][type.type_name.to_sym] if options[:scopes]
-          scope = scope.instance_eval(&additional_scope) if additional_scope
-
-          [type, scope.index_by(&:id)]
+          loaded = type.adapter.load(objects, options[type.type_name.to_sym] || {})
+          [type, loaded.index_by.with_index { |loaded, i| objects[i] }]
         end]
 
-        _results.map { |result| loaded_objects[result.class][result.id.to_i] }.compact
+        _results.map { |result| loaded_objects[result.class][result] }
       end
     end
   end